@@ -1,12 +1,20 @@
 import { createContext, useContext, useState } from "react";
 import { Order } from "../pages/Orders";
 
-const OrderContext = createContext<{order: Order | null, setOrder: React.Dispatch<React.SetStateAction<Order | null>>} | null>(null);
+type OrderContextType = {
+  order: Order | null;
+  setOrder: React.Dispatch<React.SetStateAction<Order | null>>;
+  orders: Order[] | null;
+  setOrders: React.Dispatch<React.SetStateAction<Order[] | null>>;
+};
+
+const OrderContext = createContext<OrderContextType | null>(null);
 
 export const OrderProvider: React.FC<{children: React.ReactNode}> = ({ children }) => {
     const [order, setOrder] = useState<Order | null>(null);
+    const [orders, setOrders] = useState<Order[] | null>(null);
     return (
-        <OrderContext.Provider value={{ order, setOrder }}>
+        <OrderContext.Provider value={{ order, setOrder, orders, setOrders }}>
             {children}
         </OrderContext.Provider>
     );
